reuse Layer in CanvasBase
